Caution: Run Main.m to get Model(i)Evolution.Out
Caution: It will take at most 2 minutes on a desktop and  creates 250MB
of data.  For all the models, it will create ~1.25GB of Data and takes
~5 minutes. I didn't include Model(i)Evolution.Out file here.

To get solutions to all 6 scenarios, replace the following two lines
%for K=1:6
%end %End for-loop K=1:6
with
for K=1:6
end %End for-loop K=1:6

Not sure of eiganvalues if they are correct

For tropical forest

x'=Ax+b
b is saved as Model1b.Out
A is saved as Model1Jacobi.Out
Fixed points are saved as Model1FixedPoints.Out

Model1Evolution.Out has the numerical solution to the DE that has been
calculated using Euler method with time steps of dt=0.001  for year 0 to
year 2000.

Model1Evolution.Out has 7 rows and many columns.
Column 2 has the amount of carbon in ecosystem components 1-7 at time
t=0.001.

Column i has the amount of carbon in ecosystem components 1-7 at time
t=0.001*(i-1).